RALPH: Integrate v2 and retire v1 in one commit (#89, spec §10.2)
The cutover. The app entry already rendered the v2 shell (#81 swapped App.tsx),
so what remained was the rest of §10.2 constraint 3, all in this one commit:

- v1's four surviving components die with their four test files, together, so
  the interim rule (§9.5) is satisfied exactly once with no window of red:
  Boards, CandidateList, ChartPanel and SectorTable. App.tsx's v1 body died
  earlier with the shell; Chart.tsx stays — it is the ported lightweight-charts
  seam both the sheet and the mini charts read (§10.3).
- The /api/boards alias is dropped: the route, the client's Board* type aliases
  and fetchBoards, and the fixture module's alias path. With one caller left,
  _leaders folds back into get_leaders.
- Board.tsx was reading the v1 alias (fetchBoards/BoardsResponse) rather than
  the renamed endpoint; it now reads fetchLeaders/LeadersResponse.
- openapi.json and schema.d.ts regenerate off the narrowed route table.

test_boards_seam.py's two uniquely-covered cases — the raw-return sort with the
furniture columns, and the NEW marker diffing against last session — move to
test_leaders_seam.py against /api/leaders rather than dying with the alias, so
dropping the endpoint costs no coverage.

Backend 301 passed, frontend 127 passed (the drop from 156 is the 29 v1 tests
deleted with their components), build clean.
