+++ backend/tests/test_leaders_seam.py
@@ -1,12 +1,13 @@
-"""Seam 2 (v2): the /api/leaders/{market} endpoint, and /api/boards kept as its
-alias (spec §4.4 / §10.2, issue #76).
+"""Seam 2: the /api/leaders/{market} endpoint (spec §4.4 / §10.2, issue #76).
 
 ``Boards`` is renamed **Leaders** — the name *Board* now belongs to the composite
-home screen, so leaving ``/api/boards`` behind under its old meaning is the worst
-outcome. The endpoint and its models rename to ``leaders`` / ``Leader`` /
-``LeaderRow`` / ``LeadersResponse``, and ``/api/boards`` stays an alias for the
-branch's lifetime so the whole backend lands on ``main`` ahead of any frontend
-work without breaking v1 (spec §10.2 constraint 1).
+home screen — so the endpoint and its models are ``leaders`` / ``Leader`` /
+``LeaderRow`` / ``LeadersResponse``.
+
+The boards are a read-time cut of the persisted rank table (spec §5.2): the
+endpoint reads tonight's ranks, last session's ranks (for the ``NEW`` marker) and
+each board member's bars (for the ADR column), and returns five boards of up to
+30 rows. This seam drives it end to end through TestClient.
 
 Two phase-1 leaders-row fields land here — ``sector`` and ``dollar_volume`` —
 both cheap: the read already loads exactly these names' bars for the ADR column,
@@ -48,13 +49,6 @@ def test_leaders_endpoint_returns_five_boards(seeded_store: Store):
     body = client_for(seeded_store).get("/api/leaders/IDX").json()
     assert body["market"] == "IDX"
     assert [b["lookback"] for b in body["boards"]] == list(LOOKBACKS)
-
-
-def test_boards_alias_responds_identically(seeded_store: Store):
-    # The alias is what lets the backend land on main without breaking v1: both
-    # paths serve the same session's boards, byte-for-byte (spec §10.2).
-    client = client_for(seeded_store)
-    assert client.get("/api/boards/IDX").json() == client.get("/api/leaders/IDX").json()
 
 
 def test_leaders_endpoint_empty_state_when_no_run(store: Store):
@@ -113,3 +107,49 @@ def test_phase2_fields_are_typed_null(store: Store):
     row = board["rows"][0]
     assert row["tier"] is None
     assert row["rs_pctile"] is None
+
+
+def test_leaders_rank_on_raw_return_and_carry_furniture(store: Store):
+    session = date(2026, 8, 4)
+    _publish(store, "US", session)
+    # Three names, WIN > MID > LOW on 1w; WIN also tops the 3m decile (breadth 2).
+    store.append_bars("US", "WIN", _flat_bars(140.0))
+    store.append_bars("US", "MID", _flat_bars(120.0))
+    store.append_bars("US", "LOW", _flat_bars(105.0))
+    rows = [
+        Rank("WIN", "1w", 1.0, 0.40), Rank("WIN", "3m", 0.95, 0.9),
+        Rank("MID", "1w", 0.66, 0.20),
+        Rank("LOW", "1w", 0.33, 0.05),
+    ]
+    store.append_ranks("US", session, rows)
+
+    body = client_for(store).get("/api/leaders/US").json()
+    one_w = next(b for b in body["boards"] if b["lookback"] == "1w")
+    syms = [r["symbol"] for r in one_w["rows"]]
+    assert syms == ["WIN", "MID", "LOW"]  # sorted by raw return, highest first
+
+    win = one_w["rows"][0]
+    assert win["raw_return"] == 0.40
+    assert win["breadth"] == 2  # top-decile in 1w and 3m
+    assert win["surge"] is True  # up ≥30% over the week
+    assert win["adr"] is not None  # the ADR column rides the row
+    assert one_w["rows"][1]["surge"] is False  # MID up only 20%
+
+
+def test_new_marker_diffs_against_last_session(store: Store):
+    yday, today = date(2026, 8, 3), date(2026, 8, 4)
+    _publish(store, "US", today)
+    for s in ("STAY", "FRESH"):
+        store.append_bars("US", s, _flat_bars(130.0))
+    store.append_ranks("US", yday, [Rank("STAY", "1w", 0.9, 0.5)])
+    store.append_ranks("US", today, [
+        Rank("STAY", "1w", 0.9, 0.5), Rank("FRESH", "1w", 0.8, 0.4),
+    ])
+
+    one_w = next(
+        b for b in client_for(store).get("/api/leaders/US").json()["boards"]
+        if b["lookback"] == "1w"
+    )
+    by_symbol = {r["symbol"]: r for r in one_w["rows"]}
+    assert by_symbol["STAY"]["is_new"] is False
+    assert by_symbol["FRESH"]["is_new"] is True
