--- frontend/src/CandidateList.tsx
@@ -1,119 +1,0 @@
-import { useEffect, useState } from "react";
-import { fetchCandidates, type CandidatesResponse } from "./api/client";
-
-// Both ADR columns — distance to trigger and stop width — read as a multiple of
-// one average daily range, e.g. "1.28×" (spec §5.1 / §4.6).
-function adrMultiple(value: number): string {
-  return `${value.toFixed(2)}×`;
-}
-
-/**
- * The candidate list — tonight's detections as the five-column list of the market
- * workbench (spec §5.1 / ticket 38): ticker · star score · distance to trigger ·
- * stop width in ADR · industry · k/5 breadth.
- *
- * Two properties are load-bearing:
- * - The list is **sorted by star score descending** (spec §4.7). At equal score a
- *   name that failed the envelope fit (`line_ok`) sinks below the accepted ones —
- *   a **silent** tiebreak the server applies and nothing here marks: no glyph, no
- *   column, no chart annotation. The payload carries no `line_ok` field at all.
- * - The **stop column never filters**. It highlights the affordable sub-1×ADR
- *   minority instead of marking the ~92% unaffordable majority (spec §4.6).
- *
- * The row decides whether to open the chart; the chart decides whether to trade —
- * so ADR, dollar volume, base length, the decile ranks and sector live in the
- * chart panel, not here. Clicking a row's ticker selects it via ``onSelect`` and
- * the chart panel swaps; nothing else navigates (spec §5.3).
- */
-export default function CandidateList({
-  market,
-  selected,
-  onSelect,
-}: {
-  market: string;
-  selected?: string | null;
-  onSelect?: (symbol: string) => void;
-}) {
-  const [data, setData] = useState<CandidatesResponse | null>(null);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    let live = true;
-    setData(null);
-    setError(null);
-    fetchCandidates(market)
-      .then((d) => live && setData(d))
-      .catch((e) => live && setError(String(e)));
-    return () => {
-      live = false;
-    };
-  }, [market]);
-
-  if (error)
-    return (
-      <p role="alert" className="candidate-error">
-        Could not load candidates: {error}
-      </p>
-    );
-  if (!data) return <p>Loading candidates…</p>;
-
-  if (data.candidates.length === 0)
-    return (
-      <p className="candidate-empty">
-        No candidates tonight — no name is sitting in a valid base.
-      </p>
-    );
-
-  return (
-    <section aria-label={`${market} candidates`} className="candidate-list">
-      <h2>Candidates</h2>
-      {data.ordered_by === "ticker" && (
-        <p role="note" className="sort-note">
-          Ordered by ticker — the star score sort is not yet live.
-        </p>
-      )}
-      {data.ordered_by === "score" && (
-        <p role="note" className="sort-note">
-          Sorted by star score, highest first.
-        </p>
-      )}
-      <table aria-label={`${market} candidates`}>
-        <thead>
-          <tr>
-            <th scope="col">Ticker</th>
-            <th scope="col">Star score</th>
-            <th scope="col">Distance to trigger</th>
-            <th scope="col">Stop width ÷ 1×ADR</th>
-            <th scope="col">Industry</th>
-            <th scope="col">k/5</th>
-          </tr>
-        </thead>
-        <tbody>
-          {data.candidates.map((c) => (
-            <tr key={c.symbol} aria-selected={c.symbol === selected}>
-              <th scope="row">
-                <button
-                  type="button"
-                  className="ticker-select"
-                  aria-current={c.symbol === selected}
-                  onClick={() => onSelect?.(c.symbol)}
-                >
-                  {c.symbol}
-                </button>
-              </th>
-              {/* The star score, 0–5 to one decimal; the sort key of the list. */}
-              <td className="score">{c.score.toFixed(1)}</td>
-              <td className="distance">{adrMultiple(c.dist_adr)}</td>
-              {/* Highlights the sub-1×ADR affordable minority; never a filter. */}
-              <td className={c.affordable ? "stop affordable" : "stop"}>
-                {adrMultiple(c.stopw_adr)}
-              </td>
-              <td className="industry">{c.industry ?? "—"}</td>
-              <td className="breadth">{c.breadth}/5</td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </section>
-  );
-}
